+++ src/main/java/appeng/client/gui/widgets/GuiImgButton.java
@@ -170,6 +170,10 @@ public class GuiImgButton extends GuiButton implements ITooltip {
             this.registerApp(16 * 14, Settings.PLACE_BLOCK, YesNo.YES, ButtonToolTips.BlockPlacement, ButtonToolTips.BlockPlacementYes);
             this.registerApp(16 * 14 + 1, Settings.PLACE_BLOCK, YesNo.NO, ButtonToolTips.BlockPlacement, ButtonToolTips.BlockPlacementNo);
 
+            this.registerApp(16 * 14 + 3, Settings.CPU_SELECTION_MODE, CpuSelectionMode.ANY, ButtonToolTips.CpuSelectionMode, ButtonToolTips.CpuSelectionModeAny);
+            this.registerApp(16 * 14 + 4, Settings.CPU_SELECTION_MODE, CpuSelectionMode.PLAYER_ONLY, ButtonToolTips.CpuSelectionMode, ButtonToolTips.CpuSelectionModePlayersOnly);
+            this.registerApp(16 * 14 + 5, Settings.CPU_SELECTION_MODE, CpuSelectionMode.MACHINE_ONLY, ButtonToolTips.CpuSelectionMode, ButtonToolTips.CpuSelectionModeAutomationOnly);
+
             this.registerApp(16 * 15, Settings.SCHEDULING_MODE, SchedulingMode.DEFAULT, ButtonToolTips.SchedulingMode, ButtonToolTips.SchedulingModeDefault);
             this.registerApp(16 * 15 + 1, Settings.SCHEDULING_MODE, SchedulingMode.ROUNDROBIN, ButtonToolTips.SchedulingMode, ButtonToolTips.SchedulingModeRoundRobin);
             this.registerApp(16 * 15 + 2, Settings.SCHEDULING_MODE, SchedulingMode.RANDOM, ButtonToolTips.SchedulingMode, ButtonToolTips.SchedulingModeRandom);